FreeRTOS: Fix task stack size
In FreeRTOS version V11.0.1 or earlier, if configSTACK_DEPTH_TYPE is not
defined by user in FreeRTOSConfig.h, the default value set is uint16_t,
which is unlikely to be correct.

Fixed by replacing configSTACK_DEPTH_TYPE with StackType_t, which
should always be correct as it is used internally in xTaskCreate().

See #2351.