feat: make osal an interface library
Expose a osal-interface target that is of cmake INTERFACE type.
This target only defines the osal API and allow delaying the
choice of linking of final osal port implementation to the
link stage of the binary. This allow the Osal implementation
to be defined by the integrator of used libraries.

To accomplish this:
- osal.h becomes static and common to all targets
- osal_sys.h is changed to an internal implementation
  header for the osal port
- Build files are restructured to keep build configuration of
  a port in the port directory

By default an osal cmake alias is added to remain
compatible with previous solutions.